Rework and restructure hardware tests
- Extract the RF latch-up protection and the watchdog thread into a
  separate `HardwareSetup` component. This component contains two
  targets, one which initializes and enables only the RF latch-up
  protection, and one which also feeds the watchdog. Both targets can be
  built for Linux and the COBC.
- No longer build two versions of each hardware test, one with and one
  without the watchdog. By now, we only need the ones with the watchdog.
- Move those hardware tests which do not automatically conclude but
  require human interaction to run and/or evaluate into a separate
  `ManualTests` component. Later the plan is to rework the remaining
  hardware tests to use CatchRodos and move them into the `UnitTests`.
