--- Tests/HardwareTests/RfLatchupDisablePin.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-
-namespace sts1cobcsw
-{
-auto InitializeRfLatchupDisablePins() -> void;
-auto EnableRfLatchupProtection() -> void;
-auto DisableRfLatchupProtection() -> void;
-}
