+++ Tests/ManualTests/Eps.test.cpp
@@ -1,5 +1,3 @@
-#include <Tests/HardwareTests/RfLatchupDisablePin.hpp>
-
 #include <Sts1CobcSw/Periphery/Eps.hpp>
 
 #include <rodos_no_using_namespace.h>
@@ -23,14 +21,11 @@ public:
 private:
     void init() override
     {
-        InitializeRfLatchupDisablePins();
     }
 
 
     void run() override
     {
-        EnableRfLatchupProtection();
-
         PRINTF("\nEPS test\n\n");
 
         PRINTF("\n");
