+++ Tests/ManualTests/Watchdog.test.cpp
@@ -1,5 +1,3 @@
-#include <Tests/HardwareTests/RfLatchupDisablePin.hpp>
-
 #include <Sts1CobcSw/Hal/GpioPin.hpp>
 #include <Sts1CobcSw/Hal/IoNames.hpp>
 #include <Sts1CobcSw/Utility/RodosTime.hpp>
@@ -18,14 +16,12 @@ class WatchdogTest : public RODOS::StaticThread<>
 {
     void init() override
     {
-        InitializeRfLatchupDisablePins();
         led1Gpio.Direction(hal::PinDirection::out);
     }
 
 
     void run() override
     {
-        EnableRfLatchupProtection();
         led1Gpio.Reset();
         SuspendFor(800 * ms);
         led1Gpio.Set();
